+++ backend/app/conftest.py
@@ -33,8 +33,8 @@ from Models.Aspect import AspectModel
 from Models.Rating import RatingModel
 from Models.Aspect_Rating import Aspect_Rating_Model
 from Models.UserModel import UserModel
-from Models.Role import RoleModel, Roles
-from Models.User_Role import User_Role_Model
+from Models.RoleModel import RoleModel, Roles
+from Models.UserRoleModel import UserRoleModel
 from Models.PostModel import PostModel
 ######
