+++ backend/app/main.py
@@ -17,9 +17,10 @@ from Models.Aspect import AspectModel
 from Models.Rating import RatingModel
 from Models.Aspect_Rating import Aspect_Rating_Model
 from Models.UserModel import UserModel
-from Models.Role import RoleModel
-from Models.User_Role import User_Role_Model
+from Models.RoleModel import RoleModel
+from Models.UserRoleModel import UserRoleModel
 from Models.PostModel import PostModel
+from Models.UserAccessPostModel import UserAccessPostModel
 
 
 # import database
